Support InMemoryHistory in Nubia
Summary:
Adds initial support for passing Options to Nubia object on creation. The only option that is currently supported is whether we will persist the history using `FileHistory` or not. If disabled we fallback to `InMemoryHistory`.

This also adds a new dependency on `dataclasses` which is needed until we move to Python 3.7 (it's standard in 3.7)

Reviewed By: calebmarchent

Differential Revision: D14301526

fbshipit-source-id: 9db6a6b4e3e14cdeded43e26e917b11e1a93083a