Support for prompt-toolkit>=2 (#31)
Summary:
A background breaking change to support prompt-toolkit 2+. The changes to our public interfaces are minimal. In StatusBar object we don't pass the (cli) argument anymore.

Note that prompt-toolkit 2 is not compatible with its older versions. This PR introduces the changes necessary to _start_ this support.

Key User-facing changes:
- No need to invalidate the status bar anymore. Status bar is automatically refreshed every 1 second.
- get_tokens, get_rprompt_tokens do not take cli anymore. in StatusBar.

Future work will be added to leverage the new features of prompt-toolkit, for this this is done so we can preserve all the behaviour of nubia to be the same.
Pull Request resolved: https://github.com/facebookincubator/python-nubia/pull/31

Reviewed By: ylectric

Differential Revision: D15120605

Pulled By: AhmedSoliman

fbshipit-source-id: 6cdf62fde5803b06666ed1fcd2cacb6ae7903415